+++ src/main/java/mil/nga/geopackage/GeoPackage.java
@@ -138,6 +138,16 @@ public interface GeoPackage extends GeoPackageCore {
 	public ResultSet foreignKeyCheck();
 
 	/**
+	 * Perform a foreign key check on the database table
+	 *
+	 * @param tableName
+	 *            table name
+	 * @return null if check passed, open result set with results if failed
+	 * @since 3.3.0
+	 */
+	public ResultSet foreignKeyCheck(String tableName);
+
+	/**
 	 * Perform an integrity check on the database
 	 *
 	 * @return null if check passed, open result set with results if failed